Make PCA interfaces more robust
