+++ geomstats/learning/frechet_mean.py
@@ -11,29 +11,15 @@ from sklearn.base import BaseEstimator
 
 import geomstats.backend as gs
 import geomstats.errors as error
-from geomstats.geometry.discrete_curves import ElasticMetric, SRVMetric
-from geomstats.geometry.euclidean import EuclideanMetric
 from geomstats.geometry.hypersphere import HypersphereMetric
-from geomstats.metric_geometry.graph_space import GraphSpace
-from geomstats.metric_geometry.point_set import PointSetMetric
 
-ELASTIC_METRICS = (SRVMetric, ElasticMetric)
-
-
-def _is_linear_metric(metric):
-    return isinstance(metric, EuclideanMetric)
-
-
-def _is_elastic_metric(metric):
-    return isinstance(metric, tuple(ELASTIC_METRICS))
-
-
-def _is_geodesic_metric(metric):
-    return isinstance(metric, PointSetMetric)
-
-
-def _is_graph_space(space):
-    return isinstance(space, GraphSpace)
+from ._utils import (
+    _is_elastic_metric,
+    _is_geodesic_metric,
+    _is_graph_space,
+    _is_linear_metric,
+    _warn_max_iterations,
+)
 
 
 def _scalarmul(scalar, array):
@@ -108,16 +94,7 @@ def linear_mean(points, weights=None):
 
     weighted_points = _scalarmul(weights, points)
 
-    mean = gs.sum(weighted_points, axis=0) / sum_weights
-    return mean
-
-
-def _warn_max_iterations(iteration, max_iter):
-    if iteration + 1 == max_iter:
-        logging.warning(
-            f"Maximum number of iterations {max_iter} reached. "
-            "The estimate may be inaccurate."
-        )
+    return gs.sum(weighted_points, axis=0) / sum_weights
 
 
 class BaseGradientDescent(abc.ABC):
